[chromeperf] send subscription info for triaging in skia anomalies table
The properties of 'bug component', 'bug labels' and 'bug cc emails' are needed for triaging. We have passed those for triaging in multi-graph via skia-bridge but haven't done that for the anomalies loaded to alert table.

This CL added the same logic, as in https://chromium-review.googlesource.com/c/catapult/+/5840498

Bug: b/378492826
Change-Id: If9d5ea36e9bb7de0070597efaeb8e7f93a0f9795
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6013761
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Eduardo Yap <eduardoyap@google.com>