[typ] Add in-memory artifact support
Adds support to typ's artifact and result_sink implementations for
in-memory artifacts. These artifacts are never written to disk and
instead are uploaded directly to ResultDB in the same way that stdout
and stderr are. Currently, only text artifacts of this kind are
supported.

The primary use case for this is to avoid creating many small files on
disk, which can cause Swarming tasks to be slow to finish due to needing
to clean all of them up.

Bug: chromium:376498163
Change-Id: I3ce845113a2b738bb0b00f895b4d2f631cc1f768
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6176298
Commit-Queue: John Chen <johnchen@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>