[chromeperf] filter anomalies returned to Skia by min timestamp
In the current status on Skia, while we load anomalies from chromeperf, we may
not have the actual data point for a certain anomaly backfilled on Skia. This
causes a confusion on Skia that clicking on an old anomaly does not load any
graphs.

This CL adds a minimal timestamp when loading anomalies from chromeperf to
avoid the above situation.

Bug: b/392657241
Change-Id: Ib45b2609f9bc40384349f7f90f11cf339a4f8206
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6203446
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>