[typ] Allow tests to skip outputting associated bugs
Updates typ to allow test cases to signal that associated bugs should
not be output when printing the test finish status. The primary use case
for this at the moment is to prevent typ from outputting associated bugs
when a skip is due to something other than a Skip expectation, e.g. if
the test determined that a necessary hardware feature is missing.

Previously, it would be possible to output references to bugs that are
unrelated to the test being skipped if a non-Skip expectation existed
that applied to the current test.

Bug: chromium:395919007
Change-Id: Ic4d27d1073f9a20ce2f853de88f93355f985088c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6253668
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>