Fix displaying proxy chains under the "Bad proxy" section
We updated the name of the key containing the proxy chain
information but didn't make a corresponding change to the
net log viewer:

https://chromium-review.googlesource.com/c/chromium/src/+/4987948/9/net/proxy_resolution/configured_proxy_resolution_service.cc

I tested locally and confirmed that the information is
printed as expected after making this change.

The 'use strict' and doctype changes are to satisfy presubmit
checks.

Fixed: 397701269
Change-Id: Id06f9f7c8b4c6efddce88af0805b510e43bb6a3d
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6281069
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>