Add `--starting-directory` and `--chromium-build-directory` to typ.
This adds two command line args to typ that can be used by tests.

`--starting-directory` is an optional path that can be used to
change the working directory of the process before running tests.
It can be an absolute path or a path relative to the original
working directory.

`--chromium-build-directory` is an optional path to a build dir; if
the arg is set, CHROMIUM_BUILD_DIRECTORY will be set in the test's
environment and if a test is derived from typ.TestCase, then
TestCase.chromium_build_directory will be set as well. The path
can be absolute or relative to the original working directory when
run. Note that the path is *NOT* relative to --starting-directory.

Change-Id: Ice26242b40c211a0d4f5955f5054a8b2f33c64eb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6284129
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>