[catapult] pass user from skia bisect calls
Chromeperf will call pinpoint_service to run the legacy bisect.
This is still needed because of the changes in
https://chromium-review.googlesource.com/c/catapult/+/6265711
that set alert_ids, which isn't supported in the Skia bisect call.

The calls from Skia are failing because the user field from params
is not being passed correctly. GetEmail() in utils uses
OAuth2 information, or cookies, neither which will get the email
from the SA call from Skia. Instead, we update utils to accept an
email, which is usually set in params.

Bug: b/398911040
Change-Id: I5cce48e2a10df82cf11afa077a9deb5d5c1e45e7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6362694
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>