[pinpoint] includes bot/benchmark info in gerrit update
Prior to this CL, when users submitted a patch with their Pinpoint job,
after the job finished gerrit would update with "Job Complete". When a
patch has many Pinpoint jobs associated with it, this would make it hard
to filter for relevant results.

With this CL, we now include the bot and benchmark information to help
with sorting.

Bug: chromium:404174144
Change-Id: I247da16314fc1b6b14d0836c5ff22876c58ecdb3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6373658
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>