Push zst-compressed archive instead of zip to the device
In this CL, we change the PushChangedFiles() function so that it
pushes the changed files as a zst-compressed archive instead of a zip.

The main motivation behind this change is: we want to speed things up
by extracting the archive on the device while it is being transferred
from the host to the device. This is possible because we have made the
devil_util binary extract a portion of the archive at a time.

This is not possible if we push a zip, because zip file stores some
important metadata at the end of the file, which means we have to
read the entire zip file before we start the unzipping process.

Current benchmark shows that this gives us a ~30% reduction in the
time spent in the PushChangedFiles() step.

Some followup CLs are needed:
1. Remove the zip file creation code as well as the unzip code.
2. Remove the chmod step that we currently do at the end of
PushChangedFilesZipped(). This is not necessary because we always set
the file permission to 0777 when we extract the files from the archive.
I am tempted to just do this in this CL, but I did find a previous CL
that tried to do this but got reverted because it broke some tests.
https://chromium-review.googlesource.com/c/catapult/+/4883093
So I want to play it safe and do it in a separate CL in case it gets
reverted again.
3. We can probably also remove the _ApproximateDuration() function in
device_utils.py, and always push a zst-compressed archive. If it fails,
we can fall back to pushing files individually.

No-Try: true
Bug: 397459590
Change-Id: Ia3af3998af2215a00609ba4951e8787ea93140da
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6373645
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Martin Kong <martinkong@google.com>