Make updating of device_utils_bin threadsafe
Was noticing that with local runs, where --concurrent-adb is used,
several threads try to push the file at the same time and cause
exceptions to be printed.

This also removes support for pushing device_utils_bin as a directory,
which was needed only when we used component builds, and never worked
well (we now always use static builds for it).

Bug: 397459590
Change-Id: Iac16115419badbcc0fd2d70d4abf03312acfc296
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6469273
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Haiyang Pan <hypan@google.com>
Reviewed-by: Martin Kong <martinkong@google.com>