Remove the unzip command installation code
This is a follow-up CL for
https://chromium-review.googlesource.com/6479254

The above CL replaced usages of zip files with compressed archives.
After the above CL was landed, we no longer need the unzip command
to extract zip files. As such, we are now removing the
install_commands.py which is solely responsible for installing the
unzip command to the device.

This CL is a part of two-sided patch and will be submitted alongside
https://chromium-review.googlesource.com/c/chromium/src/+/6506822

Bug: 407986583
No-Try: true
Change-Id: I51f26e14318022359bfb25f93c5f3514535dcf9e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6506403
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Martin Kong <martinkong@google.com>