[chromeperf] add culprits info in the job table
We want to correlate a benchmark/story with the files so that changing
those files may cause performance regressions on that benchmark/story.

We are looking at the Pinpoint Job table in BQ for the bisection jobs
with differences. However, while we can find the benchmark/story info
there and the differences count, we cannot see the exact commit info for
those differences.

The plan here is:
 - from the job state, extract the difference info and add a new column for those info in datastore.
 - update the export script to populate the column to BQ.

So that we can correlate the benchmark/story to a certain commit which
caused performance changes, and then the files changed in that commit.

Bug: 406405606

Change-Id: I5a9e5cc8b3b02ff5b7dc57e190f69d7da3adbb26
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6515913
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>