Escape the file paths passed to devil_util
When we hash the files, we pass the file paths to devil_util directly.
This is problematic when the file paths contain whitespaces,
because shell interprets whitespace as the end of the current argument.
This has caused test failures such as the following one:
https://ci.chromium.org/ui/p/chromium/builders/ci/android-oreo-x86-rel/14787
This CL just wraps the file paths in single quotes.

Bug: 407986583
Change-Id: I1c5d31497a4d3e54229f8fa9bde1c5be16edcda7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6517320
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Martin Kong <martinkong@google.com>