Enable streaming in _PushChangedFilesCompressedArchive()
This CL enables streaming when pushing changed files as a
compressed archive. That is, we extract the compressed archive
on the device while the archive is being transferred from the
host to the device. This is done to speed things up.
This is possible because we have specifically made the
devil_util binary extract a portion of the input at a time.

Bug: 397461053
Change-Id: I0a6c61ef4bea90100054bb40a589028a7f60684b
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6635959
Commit-Queue: Martin Kong <martinkong@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>