Add stderr to d8 RunResult
d8 binary, which is used as vinn third_party, was updated in
2fa93b2332b7a27c3066bef8d7ba9de2a5e1ab2e. After this commit
I encountered an error when running catapult tests on Windows.
The error was taken place while mapping traces using d8. The analysis
of the problem was complicated by the fact than catapult does not store
stderr in RunResult after running d8 binary. Comment in code says that
d8 does not use stderr and print errors to stdout only. But in my case
error was printed by d8 to stderr, and I was unable to discover this,
until I made some changes in catapult and vinn scripts to forward
d8 stderr.

This CL is intended to add d8 stderr to RunResult and print d8 stderr
to log when something went wrong during trace mapping.

Change-Id: Ic32afb36ab3e4688405b3e4053629a81c9380b33
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6650792
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Dmitry Shcherbinin <cher-di@yandex-team.ru>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>