Fix profile copy for Android Desktop
This change fixes an issue where the profile directory was not being
correctly copied on Android Desktop devices. This was because the
source path was not being resolved to the correct user-specific path.

This change adds a condition to resolve the source path only for PC
hardware types, which aligns with the logic for logging in as the
main user.

Bug: 418315677
Change-Id: Ibcce18976208bc767895fcb057cd3ed3f13af137
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6694790
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Auto-Submit: Darren Wu <darrenwu@chromium.org>