Devil: work around Android U WebView bug
Android 14 (U) has a bug when switching WebView provider and it may
throw a Java exception. We need to work around the exception and check
the WebView service instead.

I can't repro the bug on my test phone or emulator, so this is a
speculative fix. I tried mocking the stdout value and this seems to
handle the expected output properly.

Bug: 353572106
Test: run_py3_tests
Change-Id: Idd7049c3c56e51a36beb9f91bd509045ff1fe81b
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6847386
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>