Devil: log the current and preferred WebView package
Followup to https://crrev.com/c/6847386. The last CL did not fix the
problem, but the error output doesn't clearly indicate why and I can't
reproduce this locally. This logs the current and preferred WebView
packages in the exception message so that we can see why this is falling
through to the end of the if-else ladder.

Bug: 353572106
Test: None
Change-Id: Ia9ffe295f38e5c655237d2755a1d537544300efa
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6863223
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>