[chromeperf] cancel CQ pinpoint job if no longer needed
If the Pinpoint job is triggered by CQ, the result will be useful only
when the CQ job is still valid. If it is cancelled or failed (e.g., a new
patchset is uploaded), the Pinpoint job is no longer needed. We should cancel
those jobs to release the resources earlier.

Bug: 442629713
Change-Id: I86aab25e4ccc2089b1d3c0a333bedb62b7ddfd7e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6912327
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>