[chromeperf] add cl_number to avoid keyword exception
We added the 'cl_number' in the commit dictionary and had the value populated
in the workflow DB. Though, we did not add it to the commit model. As a result,
in the last step when we try to create a commit instance based on that dict,
'cl_number' is not recognized and thus causes exception.

This CL adds the 'cl_number' to the commit model.

Bug: 444048017
Change-Id: I0b488d65d8e63ce4dc1c151f6efed4bb4a939a06
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6943514
Reviewed-by: John Chen <johnchen@chromium.org>
Auto-Submit: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>