[chromeperf] remove transactional decorator for IsStopped()
Removing the decorator to reduce contention on datastore.

Also, marking IsStopped as a transaction does not help on fixing the
original issue b/40783051#comment9.

Consider this order:

1. Run() is called and IsStopped() returns false and the transaction
finishes.

2. Cancel() is called from another thread, and job.cancelled is set.
Then schedule.Cancel() is called and transaction finishes. So far, two
transactions make the IsStopped() and Cancel() mutually excluded, and
there is no data race.

3. Run() continues and job.put() is called. Because the job is read in
the handler even before step #1, the value of .cancelled is not aware of
#2 and eventually job.put() overrides the result of #2.

In this order, the transactional() decoration on IsStopped does not
help.

Bug: 441654952
Change-Id: I472997d324812d1f7a61a9adfc8b4a02b9d9b2a7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/6955563
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Farid (Mojtaba) Faridzad <faridzad@google.com>