Optimize the StopRecording logic in logcat monotor.
The code used to send a stop nonce to logcat, then copied the tmp record
file to the output file.
However when the device is not responding or crashed, the script will
error out at the step of sending the stop nonce due to timeout, and lead
to an empty output file, due to the tmp record file not being copied.

This CL fixes this issue by wrapping the "stop nonce" step under a try
except clause to ensure the file copying is done properly.

Bug: chromium:425720132
Change-Id: Iec1364a1252a28d9c2d4afa096eb728df9348773
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7029919
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>