[Chromium Telemetry]-Add the CDP Fetch domain to the inspector
The Fetch domain can be directly applied in the following scenario:

In some third-party pages, the data generated during the testing process
is not exposed for direct access by Telemetry. However, if the frontend
page needs to upload the test data via a network request, Telemetry can
intercept the network request for data reporting through the Fetch
domain and obtain the test data of this session.

Change-Id: Iebe4f43310231b6d55033db61643f53ec79ef2d9
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7030783
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Commit-Queue: Peng Zhou <zhoupeng.1996@bytedance.com>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>