[Dashboard] Paginate TestMetadata queries in update_test_suites
This change modifies the _ListTestSuitesAsync function to fetch TestMetadata keys in pages of 2000. This is done to prevent potential deadline exceeded errors when fetching a large number of keys.

The descriptor lookups are also parallelized within each page.

A new test case is added to verify the paging logic.

Bug: 454114985
Change-Id: I7f19fa9f1076f5c2aaf79c88beea28dddb1c9a64
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7070163
Reviewed-by: Anri Sidorov <ansid@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>