Update `DevToolsClientBackendTest` tests to exclude omnibox popup
`DevToolsClientBackendTest` queries and validates the contexts returned
from `GetUpdatedInspectableContexts()`.

A chrome feature guarded change is beginning experiments that will
increase the # of contexts in the browser by 1. It will introduce a new
context for a webUI omnibox.

When this feature is enabled, the `GetUpdatedInspectableContexts` tests
fail as there will be 1 more context than the tests are expecting.

This CL filters out this omnibox context before validating the # of
contexts.

The chromium CL to enable the feature:
https://chromium-review.googlesource.com/c/chromium/src/+/7029540

A similar catapult CL for filtering out contexts returned by
`TabListBackend.ShouldIncludeContext()`:
https://chromium-review.googlesource.com/c/catapult/+/7077077

Bug: 450029744
Change-Id: I323fdb1308d51efa374ffb8c966dc4e24470a51a
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7081434
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Auto-Submit: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>