Fix: Process test suite descriptors serially to avoid deadlocks
In , fetching and processing test suite descriptors in parallel was causing deadlock issues. This change updates the logic to process each descriptor serially within the loop.

While this may be less efficient, it is expected to resolve the deadlocks and allow the update process to complete reliably.

Bug: 454114985
Change-Id: Ie62e61d6648d72c347ef391d98a87a0140a646d8
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7078916
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>