Fix(pinpoint): Add Crossbench benchmarks to the new job fab
This change adds the Crossbench benchmarks to the list of available benchmarks in the new job fab. These benchmarks weren't previously available because they aren't scheduled on the Waterfall, so they don't get populated in the datastore. This change circumvents this limitation by hardcoding the list of Crossbench benchmarks.

It also fixes a bug where the story and story tags fields were required even when a Crossbench benchmark was selected.

Bug: 454114985
Change-Id: Ic54560c153d32916e37372454aa41a285e1a0b7c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/7103899
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Eduardo Yap <eduardoyap@google.com>